Dir. structure cleanup, ServerSetupCard component
