+++ src/components/guard/setup/ServerSetupCard/components/PickSecurityLevel.tsx
@@ -9,7 +9,7 @@ import FormErrorMessage from "components/common/FormErrorMessage"
 import { LockSimple, LockSimpleOpen } from "phosphor-react"
 import { useController, useFormContext } from "react-hook-form"
 import KeepAccessInfoText from "./KeepAccessInfoText"
-import ModeOption from "./ModeOption"
+import ModeOption from "./SecurityLevelOption"
 
 const options = [
   {
@@ -27,7 +27,7 @@ const options = [
   },
 ]
 
-const PickMode = (): JSX.Element => {
+const PickSecurityLevel = (): JSX.Element => {
   const { colorMode } = useColorMode()
 
   const {
@@ -71,4 +71,4 @@ const PickMode = (): JSX.Element => {
   )
 }
 
-export default PickMode
+export default PickSecurityLevel
