+++ src/components/guard/setup/ServerSetupCard/index.ts
@@ -1,0 +1,3 @@
+import ServerSetupCard from "./ServerSetupCard"
+
+export default ServerSetupCard
