+++ src/pages/guard/setup.tsx
@@ -1,33 +1,12 @@
-import {
-  Alert,
-  AlertDescription,
-  AlertTitle,
-  FormControl,
-  GridItem,
-  ListItem,
-  Select,
-  SimpleGrid,
-  Stack,
-  UnorderedList,
-} from "@chakra-ui/react"
-import { useWeb3React } from "@web3-react/core"
-import Button from "components/common/Button"
-import Card from "components/common/Card"
+import { GridItem, SimpleGrid } from "@chakra-ui/react"
 import CardMotionWrapper from "components/common/CardMotionWrapper"
-import FormErrorMessage from "components/common/FormErrorMessage"
 import Layout from "components/common/Layout"
-import Section from "components/common/Section"
 import DynamicDevTool from "components/create-guild/DynamicDevTool"
-import useCreate from "components/create-guild/hooks/useCreate"
-import useServerData from "components/create-guild/PickRolePlatform/components/Discord/hooks/useServerData"
-import useSetImageAndNameFromPlatformData from "components/create-guild/PickRolePlatform/hooks/useSetImageAndNameFromPlatformData"
 import DCServerCard from "components/guard/setup/DCServerCard"
-import PickMode from "components/guard/setup/PickMode"
-import { Web3Connection } from "components/_app/Web3ConnectionManager"
+import ServerSetupCard from "components/guard/setup/ServerSetupCard"
 import { AnimatePresence, AnimateSharedLayout } from "framer-motion"
-import useUploadPromise from "hooks/useUploadPromise"
 import { useRouter } from "next/router"
-import { useContext, useEffect, useMemo, useState } from "react"
+import { useEffect, useMemo, useState } from "react"
 import { FormProvider, useForm, useFormContext, useWatch } from "react-hook-form"
 import useSWR from "swr"
 
@@ -49,8 +28,6 @@ const defaultValues = {
 
 const Page = (): JSX.Element => {
   const router = useRouter()
-  const { account } = useWeb3React()
-  const { openWalletSelectorModal } = useContext(Web3Connection)
 
   const { data: servers } = useSWR("usersServers", null, {
     revalidateOnMount: false,
@@ -80,21 +57,6 @@ const Page = (): JSX.Element => {
     [selectedServer, servers]
   )
 
-  const {
-    data: { channels, serverIcon, serverName },
-  } = useServerData(selectedServer, {
-    refreshInterval: 0,
-  })
-
-  const [uploadPromise, setUploadPromise] = useState(null)
-
-  useSetImageAndNameFromPlatformData(serverIcon, serverName, setUploadPromise)
-
-  const { handleSubmit, isUploading, shouldBeLoading } = useUploadPromise(
-    methods.handleSubmit,
-    uploadPromise
-  )
-
   const [showForm, setShowForm] = useState(false)
 
   useEffect(() => {
@@ -105,26 +67,10 @@ const Page = (): JSX.Element => {
     else setShowForm(false)
   }, [selectedServer])
 
-  useEffect(() => {
-    if (!selectedServer) methods.setValue("name", "")
-    methods.setValue(
-      "name",
-      servers?.find((server) => server.value == selectedServer)?.label
-    )
-  }, [selectedServer])
-
   const resetForm = () => {
     methods.reset(defaultValues)
     methods.setValue("DISCORD.platformId", null)
   }
-
-  const { onSubmit, isLoading, response, isSigning } = useCreate()
-
-  const loadingText = useMemo((): string => {
-    if (isUploading) return "Uploading Guild image"
-    if (isSigning) return "Check your wallet"
-    return "Saving data"
-  }, [isSigning, isUploading])
 
   return (
     <Layout title={selectedServer ? "Set up Guild Guard" : "Select a server"}>
@@ -151,84 +97,7 @@ const Page = (): JSX.Element => {
 
               {showForm && (
                 <GridItem colSpan={2}>
-                  <CardMotionWrapper>
-                    <Card px={{ base: 5, sm: 6 }} py={7}>
-                      <Stack spacing={8}>
-                        <Section title="Entry channel">
-                          <FormControl
-                            isInvalid={!!methods?.formState?.errors?.channelId}
-                            isDisabled={!channels?.length}
-                            defaultValue={channels?.[0]?.id}
-                          >
-                            <Select
-                              maxW="50%"
-                              size={"lg"}
-                              {...methods?.register("channelId", {
-                                required: "This field is required.",
-                              })}
-                            >
-                              <option value={0} defaultChecked>
-                                Create a new channel for me
-                              </option>
-                              {channels?.map((channel, i) => (
-                                <option
-                                  key={channel.id}
-                                  value={channel.id}
-                                  defaultChecked={i === 0}
-                                >
-                                  {channel.name}
-                                </option>
-                              ))}
-                            </Select>
-                            <FormErrorMessage>
-                              {methods?.formState?.errors?.channelId?.message}
-                            </FormErrorMessage>
-                          </FormControl>
-                        </Section>
-                        <Section title="Security level">
-                          <PickMode />
-                        </Section>
-                        <Alert colorScheme="gray" py="3">
-                          <Stack spacing={2}>
-                            <AlertTitle>Disclaimer</AlertTitle>
-                            <AlertDescription fontSize="sm" pl="2">
-                              <UnorderedList>
-                                <ListItem>
-                                  Ethereum wallet is required for authentication
-                                </ListItem>
-                                <ListItem>
-                                  You are hiding your members and server from
-                                  unverified users
-                                </ListItem>
-                                <ListItem>
-                                  Guild Guard protects your server from bots, not
-                                  from humans with malicious intent
-                                </ListItem>
-                              </UnorderedList>
-                            </AlertDescription>
-                          </Stack>
-                        </Alert>
-                        <SimpleGrid columns={2} gap={4}>
-                          <Button
-                            colorScheme="gray"
-                            disabled={!!account}
-                            onClick={openWalletSelectorModal}
-                          >
-                            Connect wallet
-                          </Button>
-                          <Button
-                            colorScheme="green"
-                            disabled={!account}
-                            isLoading={isLoading || isSigning || shouldBeLoading}
-                            loadingText={loadingText}
-                            onClick={handleSubmit(onSubmit, console.log)}
-                          >
-                            {response ? "Success" : "Let's go!"}
-                          </Button>
-                        </SimpleGrid>
-                      </Stack>
-                    </Card>
-                  </CardMotionWrapper>
+                  <ServerSetupCard />
                 </GridItem>
               )}
             </AnimatePresence>
